+++ lib/sync/outbox/outbox_service_isolate.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:io';
 import 'dart:isolate';
-import 'dart:math';
 
 import 'package:drift/drift.dart';
 import 'package:drift/isolate.dart';
@@ -10,16 +9,13 @@ import 'package:flutter/foundation.dart';
 import 'package:lotti/blocs/sync/outbox_state.dart';
 import 'package:lotti/classes/config.dart';
 import 'package:lotti/database/common.dart';
-import 'package:lotti/database/database.dart';
 import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/database/sync_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/sync/client_runner.dart';
 import 'package:lotti/sync/encryption.dart';
 import 'package:lotti/sync/outbox/messages.dart';
-import 'package:lotti/sync/outbox_imap.dart';
-import 'package:lotti/utils/consts.dart';
-import 'package:path_provider/path_provider.dart';
+import 'package:lotti/sync/outbox/outbox_imap.dart';
 
 Future<void> entryPoint(SendPort sendPort) async {
   final port = ReceivePort();
@@ -46,12 +42,14 @@ Future<void> entryPoint(SendPort sendPort) async {
       final outbox = OutboxServiceIsolatePart(
         syncConfig: msg.syncConfig,
         networkConnected: msg.networkConnected,
+        allowInvalidCert: msg.allowInvalidCert,
+        docDir: msg.docDir,
       );
 
       unawaited(
         getIt<SyncDatabase>().watchOutboxCount().forEach((element) {
           debugPrint('db.watchOutboxCount $element');
-          outbox.restartRunner();
+          outbox.enqueueNextSendRequest();
         }),
       );
     }
@@ -62,6 +60,8 @@ class OutboxServiceIsolatePart {
   OutboxServiceIsolatePart({
     required this.syncConfig,
     required this.networkConnected,
+    required this.allowInvalidCert,
+    required this.docDir,
   }) {
     _startRunner();
   }
@@ -72,11 +72,14 @@ class OutboxServiceIsolatePart {
   final SyncDatabase _syncDatabase = getIt<SyncDatabase>();
   ImapClient? prevImapClient;
   SyncConfig syncConfig;
+  Directory docDir;
   bool networkConnected;
+  bool allowInvalidCert;
 
   void dispose() {}
 
   void _startRunner() {
+    debugPrint('OutboxServiceIsolatePart _startRunner');
     _clientRunner = ClientRunner<int>(
       callback: (event) async {
         await sendNext();
@@ -100,16 +103,6 @@ class OutboxServiceIsolatePart {
     });
   }
 
-  // Inserts a fault 25% of the time, where an exception would
-  // have to be handled, a retry intent recorded, and a retry
-  // scheduled. Improper handling of the retry would become
-  // very obvious and painful very soon.
-  String insertFault(String path) {
-    final random = Random();
-    final randomNumber = random.nextDouble();
-    return (randomNumber < 0.25) ? '${path}Nope' : path;
-  }
-
   Future<List<OutboxItem>> getNextItems() async {
     return _syncDatabase.oldestOutboxItems(10);
   }
@@ -117,25 +110,16 @@ class OutboxServiceIsolatePart {
   Future<void> sendNext() async {
     _loggingDb.captureEvent(
       'start',
-      domain: 'OUTBOX',
+      domain: 'OUTBOX_ISOLATE',
       subDomain: 'sendNext()',
     );
 
     final b64Secret = syncConfig.sharedSecret;
 
-    final enableSyncOutbox =
-        await getIt<JournalDb>().getConfigFlag(enableSyncOutboxFlag);
-
-    if (!enableSyncOutbox) {
-      _loggingDb.captureEvent(
-        'sync not enabled',
-        domain: 'OUTBOX',
-        subDomain: 'sendNext()',
-      );
-      return;
-    }
-
-    _loggingDb.captureEvent('sendNext() start', domain: 'OUTBOX');
+    _loggingDb.captureEvent(
+      'sendNext() start',
+      domain: 'OUTBOX_ISOLATE',
+    );
 
     try {
       final clientConnected = prevImapClient?.isConnected ?? false;
@@ -143,11 +127,11 @@ class OutboxServiceIsolatePart {
       _loggingDb
         ..captureEvent(
           'sendNext() networkConnected: $networkConnected ',
-          domain: 'OUTBOX',
+          domain: 'OUTBOX_ISOLATE',
         )
         ..captureEvent(
           'sendNext() clientConnected: $clientConnected ',
-          domain: 'OUTBOX',
+          domain: 'OUTBOX_ISOLATE',
         );
 
       if (!clientConnected) {
@@ -168,10 +152,9 @@ class OutboxServiceIsolatePart {
             String? encryptedFilePath;
 
             if (filePath != null) {
-              final docDir = await getApplicationDocumentsDirectory();
               final encryptedFile =
                   File('${docDir.path}${nextPending.filePath}.aes');
-              final attachment = File(insertFault('${docDir.path}$filePath'));
+              final attachment = File('${docDir.path}$filePath');
               await encryptFile(attachment, encryptedFile, b64Secret);
               encryptedFilePath = encryptedFile.path;
             }
@@ -181,6 +164,8 @@ class OutboxServiceIsolatePart {
               subject: nextPending.subject,
               encryptedMessage: encryptedMessage,
               prevImapClient: prevImapClient,
+              syncConfig: syncConfig,
+              allowInvalidCert: allowInvalidCert,
             );
             if (successfulClient != null) {
               await _syncDatabase.updateOutboxItem(
@@ -199,7 +184,10 @@ class OutboxServiceIsolatePart {
               );
             }
             prevImapClient = successfulClient;
-            _loggingDb.captureEvent('sendNext() done', domain: 'OUTBOX');
+            _loggingDb.captureEvent(
+              'sendNext() done',
+              domain: 'OUTBOX_ISOLATE',
+            );
           } catch (e) {
             await _syncDatabase.updateOutboxItem(
               OutboxCompanion(
@@ -223,7 +211,7 @@ class OutboxServiceIsolatePart {
     } catch (exception, stackTrace) {
       _loggingDb.captureException(
         exception,
-        domain: 'OUTBOX',
+        domain: 'OUTBOX_ISOLATE',
         subDomain: 'sendNext',
         stackTrace: stackTrace,
       );
@@ -237,21 +225,13 @@ class OutboxServiceIsolatePart {
   Future<void> enqueueNextSendRequest({
     Duration delay = const Duration(milliseconds: 1),
   }) async {
-    final enableSyncOutbox =
-        await getIt<JournalDb>().getConfigFlag(enableSyncOutboxFlag);
-
-    if (!enableSyncOutbox) {
-      _loggingDb.captureEvent(
-        'Sync not enabled -> not enqueued',
-        domain: 'OUTBOX',
-      );
-      return;
-    }
-
     unawaited(
       Future<void>.delayed(delay).then((_) {
         _clientRunner.enqueueRequest(DateTime.now().millisecondsSinceEpoch);
-        _loggingDb.captureEvent('enqueueRequest() done', domain: 'OUTBOX');
+        _loggingDb.captureEvent(
+          'enqueueRequest() done',
+          domain: 'OUTBOX_ISOLATE',
+        );
       }),
     );
   }