+++ lib/sync/outbox_service.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:isolate';
-import 'dart:math';
 
 import 'package:drift/drift.dart';
 import 'package:drift/isolate.dart';
@@ -18,24 +17,16 @@ import 'package:lotti/database/sync_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/sync_config_service.dart';
 import 'package:lotti/services/vector_clock_service.dart';
-import 'package:lotti/sync/client_runner.dart';
 import 'package:lotti/sync/connectivity.dart';
-import 'package:lotti/sync/encryption.dart';
 import 'package:lotti/sync/fg_bg.dart';
 import 'package:lotti/sync/outbox/messages.dart';
 import 'package:lotti/sync/outbox/outbox_service_isolate.dart';
-import 'package:lotti/sync/outbox_imap.dart';
 import 'package:lotti/utils/audio_utils.dart';
 import 'package:lotti/utils/consts.dart';
 import 'package:lotti/utils/image_utils.dart';
 import 'package:path_provider/path_provider.dart';
 
 class OutboxService {
-  OutboxService() {
-    _startRunner();
-  }
-
-  late ClientRunner<int> _clientRunner;
   final ConnectivityService _connectivityService = getIt<ConnectivityService>();
   final FgBgService _fgBgService = getIt<FgBgService>();
   final SyncConfigService _syncConfigService = getIt<SyncConfigService>();
@@ -48,22 +39,14 @@ class OutboxService {
     fgBgSubscription.cancel();
   }
 
-  void _startRunner() {
-    _clientRunner = ClientRunner<int>(
-      callback: (event) async {
-        await sendNext();
-      },
-    );
-  }
-
   void restartRunner() {
-    _clientRunner.close();
-    _startRunner();
+    // TODO: pass request to isolate
   }
 
   Future<void> startIsolate() async {
     final syncConfig = await _syncConfigService.getSyncConfig();
     final networkConnected = await _connectivityService.isConnected();
+    final docDir = await getApplicationDocumentsDirectory();
 
     final receivePort = ReceivePort();
     await Isolate.spawn(entryPoint, receivePort.sendPort);
@@ -75,6 +58,8 @@ class OutboxService {
     final loggingDbIsolate = await getIt<Future<DriftIsolate>>(
       instanceName: loggingDbFileName,
     );
+    final allowInvalidCert =
+        await getIt<JournalDb>().getConfigFlag(allowInvalidCertFlag);
 
     if (syncConfig != null) {
       sendPort.send(
@@ -83,6 +68,8 @@ class OutboxService {
           networkConnected: networkConnected,
           syncDbConnectPort: syncDbIsolate.connectPort,
           loggingDbConnectPort: loggingDbIsolate.connectPort,
+          allowInvalidCert: allowInvalidCert,
+          docDir: docDir,
         ),
       );
     }
@@ -95,208 +82,25 @@ class OutboxService {
         await getIt<JournalDb>().getConfigFlag(enableSyncOutboxFlag);
 
     if (syncConfig != null && enableSyncOutbox) {
-      await enqueueNextSendRequest();
+      debugPrint('OutboxService init $enableSyncOutbox');
+      await startIsolate();
     }
-    debugPrint('OutboxService init $enableSyncOutbox');
-
-    unawaited(startIsolate());
 
     _connectivityService.connectedStream.listen((connected) {
       if (connected) {
         restartRunner();
-        enqueueNextSendRequest();
       }
     });
 
     _fgBgService.fgBgStream.listen((foreground) {
       if (foreground) {
         restartRunner();
-        enqueueNextSendRequest();
       }
     });
-
-    Timer.periodic(const Duration(minutes: 1), (timer) async {
-      final unprocessed = await getNextItems();
-      if (unprocessed.isNotEmpty) {
-        await enqueueNextSendRequest();
-      }
-    });
-  }
-
-  // Inserts a fault 25% of the time, where an exception would
-  // have to be handled, a retry intent recorded, and a retry
-  // scheduled. Improper handling of the retry would become
-  // very obvious and painful very soon.
-  String insertFault(String path) {
-    final random = Random();
-    final randomNumber = random.nextDouble();
-    return (randomNumber < 0.25) ? '${path}Nope' : path;
   }
 
   Future<List<OutboxItem>> getNextItems() async {
     return _syncDatabase.oldestOutboxItems(10);
-  }
-
-  Future<void> sendNext() async {
-    _loggingDb.captureEvent(
-      'start',
-      domain: 'OUTBOX',
-      subDomain: 'sendNext()',
-    );
-
-    final syncConfig = await _syncConfigService.getSyncConfig();
-    final b64Secret = syncConfig?.sharedSecret;
-
-    final enableSyncOutbox =
-        await getIt<JournalDb>().getConfigFlag(enableSyncOutboxFlag);
-
-    if (!enableSyncOutbox) {
-      _loggingDb.captureEvent(
-        'sync not enabled',
-        domain: 'OUTBOX',
-        subDomain: 'sendNext()',
-      );
-      return;
-    }
-
-    if (b64Secret == null) {
-      _loggingDb.captureEvent(
-        'sync config does not exist',
-        domain: 'OUTBOX',
-        subDomain: 'sendNext()',
-      );
-      return;
-    }
-
-    _loggingDb.captureEvent('sendNext() start', domain: 'OUTBOX');
-
-    try {
-      final networkConnected = await _connectivityService.isConnected();
-      final clientConnected = prevImapClient?.isConnected ?? false;
-
-      _loggingDb
-        ..captureEvent(
-          'sendNext() networkConnected: $networkConnected ',
-          domain: 'OUTBOX',
-        )
-        ..captureEvent(
-          'sendNext() clientConnected: $clientConnected ',
-          domain: 'OUTBOX',
-        );
-
-      if (!clientConnected) {
-        prevImapClient = null;
-      }
-
-      if (networkConnected) {
-        final unprocessed = await getNextItems();
-        if (unprocessed.isNotEmpty) {
-          final nextPending = unprocessed.first;
-          try {
-            final encryptedMessage = await encryptString(
-              b64Secret: b64Secret,
-              plainText: nextPending.message,
-            );
-
-            final filePath = nextPending.filePath;
-            String? encryptedFilePath;
-
-            if (filePath != null) {
-              final docDir = await getApplicationDocumentsDirectory();
-              final encryptedFile =
-                  File('${docDir.path}${nextPending.filePath}.aes');
-              final attachment = File(insertFault('${docDir.path}$filePath'));
-              await encryptFile(attachment, encryptedFile, b64Secret);
-              encryptedFilePath = encryptedFile.path;
-            }
-
-            final successfulClient = await persistImap(
-              encryptedFilePath: encryptedFilePath,
-              subject: nextPending.subject,
-              encryptedMessage: encryptedMessage,
-              prevImapClient: prevImapClient,
-            );
-            if (successfulClient != null) {
-              await _syncDatabase.updateOutboxItem(
-                OutboxCompanion(
-                  id: Value(nextPending.id),
-                  status: Value(OutboxStatus.sent.index),
-                  updatedAt: Value(DateTime.now()),
-                ),
-              );
-              if (unprocessed.length > 1) {
-                await enqueueNextSendRequest();
-              }
-            } else {
-              await enqueueNextSendRequest(
-                delay: const Duration(seconds: 15),
-              );
-            }
-            prevImapClient = successfulClient;
-            _loggingDb.captureEvent('sendNext() done', domain: 'OUTBOX');
-          } catch (e) {
-            await _syncDatabase.updateOutboxItem(
-              OutboxCompanion(
-                id: Value(nextPending.id),
-                status: Value(
-                  nextPending.retries < 10
-                      ? OutboxStatus.pending.index
-                      : OutboxStatus.error.index,
-                ),
-                retries: Value(nextPending.retries + 1),
-                updatedAt: Value(DateTime.now()),
-              ),
-            );
-            await prevImapClient?.disconnect();
-            // ignore: unnecessary_statements
-            prevImapClient == null;
-            await enqueueNextSendRequest(delay: const Duration(seconds: 15));
-          }
-        }
-      }
-    } catch (exception, stackTrace) {
-      _loggingDb.captureException(
-        exception,
-        domain: 'OUTBOX',
-        subDomain: 'sendNext',
-        stackTrace: stackTrace,
-      );
-      await prevImapClient?.disconnect();
-      // ignore: unnecessary_statements
-      prevImapClient == null;
-      await enqueueNextSendRequest(delay: const Duration(seconds: 15));
-    }
-  }
-
-  Future<void> enqueueNextSendRequest({
-    Duration delay = const Duration(milliseconds: 1),
-  }) async {
-    final syncConfig = await _syncConfigService.getSyncConfig();
-    final enableSyncOutbox =
-        await getIt<JournalDb>().getConfigFlag(enableSyncOutboxFlag);
-
-    if (!enableSyncOutbox) {
-      _loggingDb.captureEvent(
-        'Sync not enabled -> not enqueued',
-        domain: 'OUTBOX',
-      );
-      return;
-    }
-
-    if (syncConfig == null) {
-      _loggingDb.captureEvent(
-        'Sync config missing -> not enqueued',
-        domain: 'OUTBOX',
-      );
-      return;
-    }
-
-    unawaited(
-      Future<void>.delayed(delay).then((_) {
-        _clientRunner.enqueueRequest(DateTime.now().millisecondsSinceEpoch);
-        _loggingDb.captureEvent('enqueueRequest() done', domain: 'OUTBOX');
-      }),
-    );
   }
 
   Future<void> enqueueMessage(SyncMessage syncMessage) async {
@@ -369,8 +173,6 @@ class OutboxService {
           ),
         );
       }
-
-      await enqueueNextSendRequest();
     } catch (exception, stackTrace) {
       debugPrint('enqueueMessage $exception \n$stackTrace');
       _loggingDb.captureException(