+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Enable isolate support in JournalDb, SyncDb, LoggingDb
 - Run SyncDb and LoggingDb in separate isolate (thread)
+- Run Sync outbox in isolate to avoid jank
 
 ## [0.8.169] - 2022-10-11
 ### Changed: